Forecast layout changes and new cold start data configuration

+++ components/Forecast/PreviousCalibrationRuns.vue
@@ -46,6 +46,7 @@
                   :aria-label="'Job ID ' + slotProps.data.calibration_run_id"
                   :title="'Job ID ' + slotProps.data.calibration_run_id">
                   {{ slotProps.data.calibration_run_id }}
+                  <span v-if="slotProps.data.is_locked" class="pi pi-lock"></span>
                 </span>
               </template>
             </Column>
@@ -220,14 +221,14 @@ const onRowContextMenu = (event: any) => {
     setSelectedCalibrationRunId(parseInt(event.originalEvent.currentTarget.children[0].textContent));
     cmCalibrationRun.value.push({ label: 'Run New Forecast', icon: 'pi pi-chevron-circle-right', command: () => navigateToSetupForecast() });
     cmCalibrationRun.value.push({ label: 'View Calibration Details', icon: 'pi pi-list', command: () => viewCalibrationDetails(crRowData.calibration_run_id) })
-    //cmCalibrationRun.value.push( { label: 'Evaluate', icon: 'pi pi-chart-scatter', command: () => openSelectedCalibrationRun() } );
-    //cmCalibrationRun.value.push( { label: 'Show Setup', icon: 'pi pi-list', command: () => onCalibrationRunForForecastRowSelect() } );
     if (calibrationRunForForecast.value?.is_downloadable) {
       cmCalibrationRun.value.push({ label: 'Download Results', icon: 'pi pi-download', command: () => downloadSelectedCalibrationData() });
     }
     cmCalibrationRun.value.push({ label: 'Export Calibration Config', icon: 'pi pi-file-export', command: () => exportSelectedCalibrationData() });
-    cmCalibrationRun.value.push({ label: 'Delete', icon: 'pi pi-trash', command: () => deleteSelectedCalibrationRun() });
-    if (!selectedCalibrationRun.value?.is_archived) {
+    if (!calibrationRunForForecast?.value?.status.includes('Submitted') && !calibrationRunForForecast?.value?.status.includes('Running') && !calibrationRunForForecast?.value?.is_locked) {
+      cmCalibrationRun.value.push({ label: 'Delete', icon: 'pi pi-trash', command: () => deleteSelectedCalibrationRun() });
+    }
+    if (!calibrationRunForForecast?.value?.status.includes('Submitted') && !calibrationRunForForecast?.value?.status.includes('Running') && !calibrationRunForForecast.value?.is_archived) {
       cmCalibrationRun.value.push({ label: 'Archive', icon: 'pi pi-folder', command: () => archiveSelectedCalibrationRun(true) });
     }
   }
@@ -375,9 +376,28 @@ const deleteSelectedCalibrationRun = () => {
 const acceptDelete = (selectedRunId: number) => {
   deleteCalibrationRun(selectedRunId).then(response => {
     if (response.status === 200) {
-      const tMsg: ToastMessageOptions = { severity: useApiResponseToastSeverityCode(response?.status), 
-      summary: 'Delete Calibration Run', detail: 'Job ' + selectedRunId + ' deleted', life: useApiResponseToastSeverityLife(response?.status)};
-      toast.add(tMsg); addToastRecord(tMsg);   
+      let successMessages: string[] = [];
+      let failureMessages: string[] = [];
+      response._data?.jobs.forEach(job => {
+        if (job.success) {
+          successMessages.push(job.message);
+        } else {
+          failureMessages.push(job.message);
+        }
+      });
+      toast.removeAllGroups();
+      if (successMessages.length > 0) {
+        const tMsg: ToastMessageOptions = { severity: 'success', 
+          summary: 'Delete Job', detail: successMessages.join('\n'), 
+          life: ToastTimeout.timeoutSuccess};
+        toast.add(tMsg); addToastRecord(tMsg);
+      }
+      if (failureMessages.length > 0) {
+        const tMsg: ToastMessageOptions = { severity: 'error', 
+          summary: 'Delete Job', detail: failureMessages.join('\n'), 
+          life: ToastTimeout.timeoutError};
+        toast.add(tMsg); addToastRecord(tMsg);
+      } 
       getCalibrationJobsForForecast();
     } else {
       useApiErrorResponsePreprocess(response).forEach(message => {
@@ -414,9 +434,28 @@ const archiveSelectedCalibrationRun = () => {
 const acceptArchive = (selectedRunId: number) => {
   archiveCalibrationRun(selectedRunId, true).then(async (response) => {
     if (response.status === 200) {
-      const tMsg: ToastMessageOptions = { severity: 'success', 
-        summary: 'Calibration Job Archived', detail: 'Job ' + selectedRunId + ' Archived', life: ToastTimeout.timeoutSuccess };
-      toast.add(tMsg); addToastRecord(tMsg);
+      let successMessages: string[] = [];
+      let failureMessages: string[] = [];
+      response._data?.jobs.forEach(job => {
+        if (job.success) {
+          successMessages.push(job.message);
+        } else {
+          failureMessages.push(job.message);
+        }
+      });
+      toast.removeAllGroups();
+      if (successMessages.length > 0) {
+        const tMsg: ToastMessageOptions = { severity: 'success', 
+          summary: 'Archive Job', detail: successMessages.join('\n'), 
+          life: ToastTimeout.timeoutSuccess};
+        toast.add(tMsg); addToastRecord(tMsg);
+      }
+      if (failureMessages.length > 0) {
+        const tMsg: ToastMessageOptions = { severity: 'error', 
+          summary: 'Archive Job', detail: failureMessages.join('\n'), 
+          life: ToastTimeout.timeoutError};
+        toast.add(tMsg); addToastRecord(tMsg);
+      }
       getCalibrationJobsForForecast();
     } else {
       useApiErrorResponsePreprocess(response).forEach(message => {
